Project7
Assigned a local and global variables. Wrote a function that includes an if statement. Intentionally wrote a function with an error in it and used the consol.log() method to debug it within the console in the Chrome Dev Tools. Was provided a function two add to the JS and HTML file that will display the time of day by Morning, Afternoon or evening reference.