Project10
Wrote a function that utilizes a while loop, wrote a function that utilizes a for loop, wrote a function that utilizes an array, created an object using the "let" keyword,
Wrote HTML elements for  each JS function to be called and displayed from the JavaScript code. Commented throughout the HTML and JS file to explain how the code works.